+++ src/components/App/App.module.css
@@ -1,7 +1,6 @@
 .app {
-  height: 100%;
   text-align: center;
-  max-width: 1400px;
-  overflow-x: hidden;
+  max-width: 80rem;
+  overflow: hidden;
   margin: auto;
 }
