+++ src/components/Base/Card/Index.tsx
@@ -1,0 +1,1 @@
+export {}
