+++ src/components/Hero/Hero.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
 import { Hero } from './Hero';
-import { Section } from '../Section';
+import { Section } from '../Base/Section';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
