+++ src/index.css
@@ -1,5 +1,4 @@
 body {
-  height: 100%;
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans',
     'Droid Sans', 'Helvetica Neue', sans-serif;
@@ -7,13 +6,13 @@ body {
   -moz-osx-font-smoothing: grayscale;
 }
 
-html {
+/* html {
   height: 100%;
 }
 
 #root {
   height: 100%;
-}
+} */
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
